API設計書に基づくリファクタリング(もくもく会一覧部分)
